Check data_list is in correct formats

When new checks come up, add them in the same format: logical tests first and then add them to the if statement

data_list_correct(raw_json)

Arguments

raw_json
Raw JSON response from datos.gob.es

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.