Extract data and metadata from a given data set of https://datos.gob.es/

Extract data and metadata from a given data set of https://datos.gob.es/

openes_load(x, encoding = "UTF-8", guess_encoding = TRUE, ...)

Arguments

x
A tibble given by openes_keywords only containing one dataset (1 row) or the end path of a dataset such as 'l01280148-seguridad-ciudadana-actuaciones-de-seccion-del-menor-en-educacion-vial-20141' from https://datos.gob.es/es/catalogo/l01280148-seguridad-ciudadana-actuaciones-de-seccion-del-menor-en-educacion-vial-20141.

encoding
The encoding passed to read (all) the files. Most cases should be resolved with either 'UTF-8', 'latin1' or 'ASCII'.

guess_encoding
A logical stating whether to guess the encoding. This is set to TRUE by default. Whenever guess_encoding is set to TRUE, the 'encoding' argument is ignored. If guess_encoding fails to guess the encoding, openes_load falls back to the encoding argument.

...
Arguments passed to read_csv and the other related read_* functions from readr. Internally, openes_load determines the delimiter of the file being read but the arguments for each of these functions are practically the same, so it doesn't matter how openes_load determines the delimiter, any of the arguments will work on all read_* functions.

Value

if path_id is a valid dataset path, a list with two slots: metadata and data. Each slot contains tibble's that contain either metadata or the data itself. If path_id is not a valid dataset path, it returns an empty list. See the details section for some caveats.

Details

openes_load can return two possible outcomes: either an empty list or a list with a slot called metadata and another slot called data. Whenever the path_id argument is an invalid dataset path, it will return an empty list. When path_id is a valid dataset path, openes_load will return an a list with the two slots described above.

For the metadata slot, openes_load returns a tibble with most available metadata of the dataset. The columns are:

keywords: the available keywords from the dataset in the homepage of the dataset.

language: the available languages of the dataset's metadata. Note that that this does not mean that the dataset is in different languages but only the metadata.

description: a short description of the data being read.

url: the complete url of the dataset in https://datos.gob.es/. Note that this URL is not the access URL to the dataset but to the dataset's homepage in https://datos.gob.es/.

date_issued: the date at which the dataset was uploaded.

date_modified: the date at which the last dataset was uploaded. If the dataset has only been uploaded once, this will return 'No modification date available'.

publisher: the entity that publishes the dataset. See openes_load_publishers for all available publishers.

publisher_data_url: the homepage of the dataset in the website of the publisher. This is helpful to look at the definitions of the columns in the dataset.

The metadata of the API can sometimes be returned in an incorrect order. For example, there are cases when there are several languages available and the order of the different descriptions are not in the same order of the languages. If you find any of these errors, try raising the issue directly to https://datos.gob.es/ as the package extracts all metadata in the same order as it is.

Whenever the metadata is in different languages, the resulting tibble will have the same numer of rows as there are languages containing the different texts in different languages and repeating the same information whenever it's similar across languages (such as the dates, which are language agnostic).

In case the API returns empty requests, both data and metadata will be empty tibble's with the same column names.

For the data slot, openes_load returns a list containing at least one tibble. If the dataset being request has file formats that openes_load can read (see permitted_formats) it will read those files. If that dataset has several files, then it will return a list of the same length as there are datasets where each slot in that list is a tibble with the data. If for some reason any of the datasets being read cannot be read, openes_load has a fall back mechanism that returns the format that attempted to read together with the URL so that the user can try to read the dataset directly. In any case, the result will always be a list with tibble's where each one could be the requested dataset (success) or a dataset with the format and url that attempted to read but failed (failure).

Inside the data slot, each list slot containing tibble's will be named according to the dataset that was read. When there is more than one dataset, the user can then enter the website in the url column in the metadata slot to see all names of the datasets. This is handy, for example, when the same dataset is repeated across time and we want to figure out which data is which from the slot.

The API of https://datos.gob.es/ is not completely homogenous because it is an aggregator of many different API's from different cities and provinces of Spain. openes_load can only read a limited number of file formats but will keep increasing as the package evolves. You can check the available file formats in permitted_formats. If the file format of the requested path_id is not readable, openes_load will return a list with only one tibble with all available formats with their respective data URL inside the data slot so that users can read the manually.

In a similar line, in order for openes_load to provide the safest behavior, it is very conservative in which publisher it can read from https://datos.gob.es/. Because some publishers do not have standardized datasets, reading many different publishers can become very messy. openes_load currently reads files from selected publishers because they offer standardized datasets which makes it safer to read. As the package evolves and the data quality improves between publishers, the package will include more publishers. See the publishers that the package can read in publishers_available.

Examples

# For a dataset with only one file to read
example_id <- 'l01080193-fecundidad-madres-de-15-a-19-anos-quinquenal-2003-2014'
some_data <- openes_load(example_id)

# Print the file to get some useful information
some_data

#> <datos.gob.es API>
#>    Description: Fecundidad madres de 15 a 19 años (quinquenal). 2003-2014
#>    Publisher: Ayuntamiento de Barcelona
#>    Languages: es, ca, en
#>    Date of release: NA
#>    # of files read: 1 out of 1

# Access the metadata
some_data$metadata

#> # A tibble: 3 x 8
#>   keywords language description url   date_issued         date_modified
#>   <chr>    <chr>    <chr>       <chr> <dttm>              <chr>        
#> 1 Barrios~ es       Fecundidad~ http~ NA                  No modificat~
#> 2 Barrios~ ca       Fecunditat~ http~ NA                  No modificat~
#> 3 Barrios~ en       Fertility ~ http~ NA                  No modificat~
#> # ... with 2 more variables: publisher <chr>, publisher_data_url <chr>

# Access the data. Note that the name of the dataset is in the list slot. Whenever
# there are different files being read, you might want to enter to homepage
# of the dataset in datos.gob.es with some_data$metadata$url or directly
# to the homepage of dataset at the publisher's website
# some_data$metadata$publisher_data_url
some_data$data

#> $t02.csv
#> # A tibble: 75 x 11
#>    Dte.  Barris `2003-2007` `2004-2008` `2005-2009` `2006-2010` `2007-2011`
#>    <chr> <chr>        <dbl>       <dbl>       <dbl>       <dbl>       <dbl>
#>  1 BARC~ NA            1359        1381        1404        1363        1310
#>  2 1     1. el~          76          77          80          78          79
#>  3 1     2. el~          19          16          13          10           9
#>  4 1     3. la~          15          14          17          18          15
#>  5 1     4. Sa~          24          27          23          24          25
#>  6 2     5. el~          15          11          15          12          10
#>  7 2     6. la~          28          20          17          16          16
#>  8 2     7. la~          10          10           7           6           5
#>  9 2     8. l'~          13          17          16          14          14
#> 10 2     9. la~          19          21          21          18          18
#> # ... with 65 more rows, and 4 more variables: `2008-2012` <dbl>,
#> #   `2009-2013` <dbl>, `2010-2014` <dbl>, `2011-2015` <dbl>
#> 

# For a dataset with many files

# NOT RUN {
example_id <- 'l01080193-domicilios-segun-nacionalidad'
res <- openes_load(example_id)

# Note that you can see how many files were read in '# of files read'
res

# See how all datasets were read but we're not sure what each one means.
# Check the metadata and read the description. If that doesn't do it,
# go to the URL of the dataset from the metadata.
res$data

# Also note that some of the datasets were not read uniformly correct. For example,
# some of these datasets were read with more columns or more rows. This is left
# to the user to fix. We could've added new arguments to the `...` but that would
# apply to ALL datasets and it then becomes too complicated.

# Encoding problems

long <- "l01080193-descripcion-de-la-causalidad-de-los-accidentes"
string <- "-gestionados-por-la-guardia-urbana-en-la-ciudad-de-barcelona"

id <- paste0(long, string)
pl <- openes_load(id)

# The dataset is read successfully but once we print them, there's an error
pl$data
$`2011_ACCIDENTS_CAUSES_GU_BCN_.csv`
Error in nchar(x[is_na], type = "width") :
  invalid multibyte string, element 1

# This error is due to an encoding problem.
# We can use readr::guess_encoding to determine the encoding and reread

# This suggests an ASCII encoding
library(readr)
guess_encoding(pl$data[[1]])

pl <- openes_load(id, 'ASCII')

# Success
pl$data

# For exploring datasets with openes_keywords and piping to openes_load
library(dplyr)

kw <- openes_keywords("turismo", "l01080193") # Ayuntamiento de Barcelona#'
kw

dts <-
 kw %>%
 filter(is_readable == TRUE,
        grepl("Tipos de propietarios", description)) %>% # Narrow it down to only 1 dataset
 openes_load()

dts$metadata

dts$data
# }